mattdrayer/cert-webview-add-org-fields: Added new fields to context
